.	Added the timestamp in the video description. .	Now user table cells are clickable. .	Moved the filter options from navigation bar to view like user settings. .	Now user can watch the video on original source in UIWebView. .	Moved settings button only to logged in user profile view. .	Deleted labels folder from ui folder. .	Fixed the keyboard error for user profile settings view. .	Added the controls to full screen mode of video player. .
git-svn-id: https://mykikin.jira.com/svn/KM/trunk/ios/KikinVideo@24033 e6edf6fb-f266-4316-afb4-e53d95876a76

+++ Classes/objects/ActivityObject.h
@@ -27,14 +27,14 @@
 //                    User activity object
 // -----------------------------------------------------------
 @interface ActivityObject : NSObject {
-	long                timestamp;
+	double                timestamp;
     NSString*           activity_heading;
     NSArray*            activityHeadingLabelsList;
     NSMutableArray*     userActivities;
     VideoObject*        video;
 }
 
-@property()			long timestamp;
+@property()			double timestamp;
 @property(retain)	NSString* activity_heading;
 @property(retain)	NSArray* activityHeadingLabelsList;
 @property(retain)	NSMutableArray* userActivities;
